requirements.txt
asgiref==3.9.1
Django==5.2.5
sqlparse==0.5.3
tzdata==2025.2
absl-py==2.3.1
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
flatbuffers==25.2.10
fonttools==4.59.1
humanfriendly==10.0
idna==3.10
jax==0.7.1
jaxlib==0.7.1
kiwisolver==1.4.9
matplotlib==3.10.5
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.22.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
scipy==1.16.1
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
sqlparse==0.5.3
sympy==1.14.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
